[aaa] reorg migration module
