+++ pallets/ajuna-awesome-avatars/src/migration/mod.rs
@@ -14,8 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod v6;
-mod weights;
+mod v6;
 
 use super::*;
 use frame_support::traits::OnRuntimeUpgrade;
