+++ pallets/ajuna-awesome-avatars/src/migration/v5/v5.rs
