+++ pallets/ajuna-awesome-avatars/src/migration/v6/weights.rs
